speed up golden rendering
